world: Mesh chunks globally
Chunks are now meshed using the whole world state instead of just the
section. This is less performant but fixes the ugly ambient occlusion seams for
now.

Additionally, upon setting a block, depending on where the block was
placed, neighboring chunks touching that block will now be marked dirty
accordingly. In the future this should also be done to chunk updates.